// lib/plant_tracker/model/activityOut.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Book Store API
 * Book Store API for renting books and notifying available/returned books in a store
 * OpenAPI spec version: 1.0.0
 */
import type { ActivityOutDescription } from "./activityOutDescription";

export interface ActivityOut {
  description?: ActivityOutDescription;
  id?: string;
  name: string;
}