// lib/plant_tracker/model/plantOut.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Plant Tracker API
 * API for interacting with the Plant Tracker application
 * OpenAPI spec version: 1.0.0
 */
import type { PlantOutCommonName } from "./plantOutCommonName";
import type { PlantOutDeathDate } from "./plantOutDeathDate";
import type { PlantOutMainPhoto } from "./plantOutMainPhoto";
import type { PlantOutNotes } from "./plantOutNotes";
import type { PlantOutPurchaseDate } from "./plantOutPurchaseDate";
import type { PlantOutScientificName } from "./plantOutScientificName";

export interface PlantOut {
  area: string;
  common_name?: PlantOutCommonName;
  death_date?: PlantOutDeathDate;
  graveyard?: boolean;
  id?: string;
  main_photo?: PlantOutMainPhoto;
  name: string;
  notes?: PlantOutNotes;
  purchase_date?: PlantOutPurchaseDate;
  scientific_name?: PlantOutScientificName;
  user: string;
}